Unknown composition materials (#587)
* work

* Convert unknown composition simplefluids

flop
The dealing of the first three face-up cards to the board, refers also to those three cards themselves. Also see turn and river.

* Finished unknown composition materials

Big Shlyopa (also known simply as Shlyopa or Big Russian Cat) is a series of Internet memes with a caracal named Gosha. In the English-speaking segment of the Internet, the caracal is known as Floppa or Big Floppa.

* Move Polyimide and FluorinatedEthylenePropylene to first degree materials